refactor(backend): rename auth module to authn
Pair the authentication module with the new authz (authorization) module:
`crate::auth` → `crate::authn` (jwt, password, role, middleware, bootstrap).
Only the module path changes — identifiers like AuthUser, auth_user, and
auth_required keep their names. The signup/login handler file
(crate::handlers::auth) is a separate namespace and is left as-is.
