+++ backend/src/handlers/automation.rs
@@ -13,7 +13,7 @@ use sha2::{Digest, Sha256};
 use uuid::Uuid;
 
 use crate::{
-    auth::{AuthUser, Role},
+    authn::{AuthUser, Role},
     error::{ApiResult, AppError},
     model::{
         AutomationListResponse, AutomationResponse, CreateAutomationRequest, CreateTriggerRequest,
