+++ backend/src/handlers/project.rs
@@ -8,7 +8,7 @@ use axum::{
 use uuid::Uuid;
 
 use crate::{
-    auth::AuthUser,
+    authn::AuthUser,
     error::{ApiResult, AppError},
     events::WsEvent,
     handlers::session::cleanup_session_resources,
